Custom User Improvements
Fixed a bug where a users last name was treated as a required parameter
by the user creation method, even though it should be an optional field.

Modified CustomUser model definition to allow last_name column to be
nullable and changed default value of is_active to True.

Added tests for the creation of normal users and superusers.

Signed-off-by: Avery Uslaner <ctsadmin@conservationtechnologysolutions.org>